requirements.txt
qdrant-client
sentence-transformers
pdfminer.six
python-markdown
python-dotenv